[Fix] CloudKitService 파일 정리
