+++ Dorocat/Data/2. Service/4. CloudKitService/Model/3. CKError.swift
@@ -7,6 +7,3 @@
 
 import Foundation
 
-enum CKError: Error {
-    
-}
